backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
anthropic==0.26.1
python-dotenv==1.0.0

